pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "manager"
version = "0.0.0"
[tool.poetry.dependencies]
beautifulsoup4 = "*"
firebase-admin = "*"
flask = "*"
gcloud = "*"
google = "*"
google-api-core = "*"
google-api-python-client = "*"
google-auth = "*"
google-auth-httplib2 = "*"
google-cloud-core = "*"
google-cloud-firestore = "*"
google-cloud-storage = "*"
google-resumable-media = "*"
googleapis-common-protos = "*"
pillow = "*"
pyrebase = "*"
python = "^3.7"
requests-toolbelt = "*"
robobrowser = "*"
virtualenv = "*"
virtualenv-clone = "*"
jwt = "^0.6.1"
requests = "^2.22"
